DelayCalculatorTest moved to right package
