+++ import/src/main/java/com/google/cloud/healthcare/imaging/dicomadapter/ImportAdapter.java
@@ -116,9 +116,12 @@ public class ImportAdapter {
     Map<DestinationFilter, IDicomWebClient> destinationMap = configureDestinationMap(
         flags.destinationConfigInline, flags.destinationConfigPath, credentials);
 
+    // backup upload service
     IBackupUploader backupUploader = new LocalBackupUploader();
     IBackupUploadService backupUploadService = new BackupUploadService(uploadPath,
         backupUploader, new DelayCalculator(uploadRetryAmount, minUploadDelay, maxWaitingTimeBtwUploads));
+    //todo: parse --persistent_file_storage_location -> create backupUploadService instance - if present.
+    // else - null.
 
     DicomRedactor redactor = configureRedactor(flags);
     CStoreService cStoreService =
